Work-in-progress on fork.
Specify the http://loinc.org| system for all Observation token based searches.
This is necessary due to the current implementation of the Cerner FHIR server.

Specify requested scopes explicitly rather than using wildcard scopes.
This is necessary due to the current implementation of the Cerner Authorization server.

Fix the trend message when there is only a single growth assessment.

All of this work will be contributed back to the open source upstream project.